+++ docs/CHANGELOG.md
@@ -20,6 +20,7 @@ General:
    - General optimizations
    - Added a new voxelization algorithm and a cvar to use it (set `voxformat_voxelizemode` to 1)
    - Fixed issues with importing images with a depth-map
+   - Fixed regression for reading minecraft voxel data
 
 VoxEdit: